requirements.txt
gender-guesser==0.4.0
pyodbc==5.3.0
python-dotenv==1.2.2
